Dev changes for 1.2.0 (#25)
* IRV now handles missing values, and psychsyn handles NA results better (#17)

* typos and renamed carelessDataset columns

- typo in IRV / psacing in psychsyn
- carelessDataset now has cogent column names

* changed vignettes

removed introduction vignette from package (now hosted online) and fixed typo with linking vignette

* changed psychsyn to do a resampling

psychsyn comes up with different correlation coefficients dependent on whether the item of an item pair is located at the x or y column.

* psychsyn now resamples

psychsyn correlation values was dependent on the placement of items of item pairs in the x and y columns. psychsyn now does multiple random (re)placements of the items.

* irv now can handle missing values

* Updated psychsyn to do up to 10 resamples while encountering NAs

In response to Issue #16

* updated IRV to make na.rm optional

* updating docs to match francisco's changes

* remove vignette

* adding suggestion for texlive-fonts-extra

* change to travis-ci config to attempt to account for latex issues

* trying a different before install command

* adding one more

* changing repo for texlive

* trying http mirror instead

* try again

* idfk

* still not working

* another try

* one more thing before bed

* macos why?!?

* again

* trying to work around.

* no cache

* last try

* travis is not allowed to macos anymore

* feature updates for psychsyn, mahad, irv (#18)

* typos and renamed carelessDataset columns

- typo in IRV / psacing in psychsyn
- carelessDataset now has cogent column names

* changed vignettes

removed introduction vignette from package (now hosted online) and fixed typo with linking vignette

* changed psychsyn to do a resampling

psychsyn comes up with different correlation coefficients dependent on whether the item of an item pair is located at the x or y column.

* psychsyn now resamples

psychsyn correlation values was dependent on the placement of items of item pairs in the x and y columns. psychsyn now does multiple random (re)placements of the items.

* irv now can handle missing values

* Updated psychsyn to do up to 10 resamples while encountering NAs

In response to Issue #16

* updated IRV to make na.rm optional

* updated mahad to handle NA properly

* Update mahad.R - fixed typo

fixed typo

* Update .gitignore

* Updates to mahaD, psychsyn (#21)

* typos and renamed carelessDataset columns

- typo in IRV / psacing in psychsyn
- carelessDataset now has cogent column names

* changed vignettes

removed introduction vignette from package (now hosted online) and fixed typo with linking vignette

* changed psychsyn to do a resampling

psychsyn comes up with different correlation coefficients dependent on whether the item of an item pair is located at the x or y column.

* psychsyn now resamples

psychsyn correlation values was dependent on the placement of items of item pairs in the x and y columns. psychsyn now does multiple random (re)placements of the items.

* irv now can handle missing values

* Updated psychsyn to do up to 10 resamples while encountering NAs

In response to Issue #16

* updated IRV to make na.rm optional

* updated mahad to handle NA properly

* Update mahad.R - fixed typo

fixed typo

* Update .gitignore

* updated psychsyn

handling of NA improved. Added corresponding test.

Co-authored-by: fwilhelm <franciscowilhelm@users.noreply.github.com>
Co-authored-by: Richard Yentes <ryentes@gmail.com>

* changed psychsyn to default resample

* Changed evenodd so it's interpreted similarly to other metrics addresses #19

* fixes to psychsyn doc for resampling

* update travis.yml to skip oldrel

* Cleaning up dev

* Update .travis.yml

* removing travis ci

* Update DESCRIPTION

* updating docs

* changes to even-odd (#23)

* changes to even-odd

1. Introduced error when calling even odd with just one factor.
2. Introduced warning when variables indicated by factor argument do not match the input data (x).
3. Use own warning for cases when NA arises because of 0 variance in even or odd vectors.
4. refactored coding using apply.

* evenodd minor change

added an explicit error when factors > x. Adjusted the call. parameter to be FALSE.

* added warning to even-odd that computation has changed (#24)

Co-authored-by: Francisco Wilhelm <38857150+franciscowilhelm@users.noreply.github.com>
Co-authored-by: fwilhelm <franciscowilhelm@users.noreply.github.com>